Add command to create pull requests
Pull requests need to have at least one commit
different from the base branch. Therefore, they
cannot be created when a branch is immediately
created for a ticket because this happens on the
base branch. This means that a separate command is
needed to create new pull requests that can be run
once there have been some commits.